Refactor small climber actions
Refactored the package structure to separate actions (small things) and sequences (groups of actions).
