+++ src/main/java/ca/team2706/frc/robot/OI.java
@@ -1,7 +1,7 @@
 package ca.team2706.frc.robot;
 
-import ca.team2706.frc.robot.commands.climber.Climb;
-import ca.team2706.frc.robot.commands.climber.RetractClimberPistons;
+import ca.team2706.frc.robot.commands.climber.sequences.Climb;
+import ca.team2706.frc.robot.commands.climber.actions.RetractClimberPistons;
 import ca.team2706.frc.robot.commands.drivebase.AbsoluteRotateWithGyro;
 import ca.team2706.frc.robot.commands.drivebase.CurvatureDriveWithJoystick;
 import ca.team2706.frc.robot.commands.drivebase.DriverAssistVision;
