+++ src/main/java/ca/team2706/frc/robot/commands/climber/actions/RetractClimberMotors.java
@@ -1,4 +1,4 @@
-package ca.team2706.frc.robot.commands.climber;
+package ca.team2706.frc.robot.commands.climber.actions;
 
 import ca.team2706.frc.robot.subsystems.ClimberMotor;
 import edu.wpi.first.wpilibj.command.Command;
