+++ src/main/java/ca/team2706/frc/robot/commands/climber/actions/RunClimberMotor.java
@@ -1,4 +1,4 @@
-package ca.team2706.frc.robot.commands.climber;
+package ca.team2706.frc.robot.commands.climber.actions;
 
 import ca.team2706.frc.robot.subsystems.ClimberMotor;
 import ca.team2706.frc.robot.subsystems.Pneumatics;
